6주차 대규모 트래픽&데이터 처리 기본 과제 수행 (Distributed Lock)

+++ hhplus-server-java/src/main/java/kr/hhplus/be/server/application/order/OrderCommand.java
@@ -8,7 +8,7 @@ import java.util.Map;
 @Builder
 public record OrderCommand (
         Long userId,
-        Map<Long, Long> productGrp,
+        Map<Long, Integer> productGrp,
         Long couponId,
         OrderAddressInfo orderAddressInfo
 ) {
